+++ bastion/provider.go
@@ -8,14 +8,16 @@ import (
 )
 
 const (
-	versionValidate33 = "v3.3"
-	versionValidate36 = "v3.6"
+	VersionWallixAPI33 = "v3.3"
+	VersionWallixAPI36 = "v3.6"
+	VersionWallixAPI38 = "v3.8"
 )
 
 func defaultVersionsValid() []string {
 	return []string{
-		versionValidate33,
-		versionValidate36,
+		VersionWallixAPI33,
+		VersionWallixAPI36,
+		VersionWallixAPI38,
 	}
 }
 
@@ -46,7 +48,7 @@ func Provider() *schema.Provider {
 			"api_version": {
 				Type:        schema.TypeString,
 				Optional:    true,
-				DefaultFunc: schema.EnvDefaultFunc("WALLIX_BASTION_API_VERSION", "v3.3"),
+				DefaultFunc: schema.EnvDefaultFunc("WALLIX_BASTION_API_VERSION", VersionWallixAPI33),
 			},
 		},
 		DataSourcesMap: map[string]*schema.Resource{
@@ -72,6 +74,7 @@ func Provider() *schema.Provider {
 			"wallix-bastion_externalauth_kerberos":                 resourceExternalAuthKerberos(),
 			"wallix-bastion_externalauth_ldap":                     resourceExternalAuthLdap(),
 			"wallix-bastion_externalauth_radius":                   resourceExternalAuthRadius(),
+			"wallix-bastion_externalauth_saml":                     resourceExternalAuthSaml(),
 			"wallix-bastion_externalauth_tacacs":                   resourceExternalAuthTacacs(),
 			"wallix-bastion_ldapdomain":                            resourceLdapDomain(),
 			"wallix-bastion_ldapmapping":                           resourceLdapMapping(),